feat: implement EventMapCard component
